Make run ids have no gaps. Also introduce schema versioning.
Run id gaps, besides being ugly, make some things harder to do. So
instead of having a serial there (which gets bumped even on failed
transactions), use MAX(id)+1 for the next id.

To have a consistent database, this needs a migration, which is
implemented as well. To ease future migrations, introduce a new table
which contains the schema version.